Refactor: Separate Play Integrity and Key Attestation modules
I have separated the Play Integrity and Key Attestation features from the `:ui:main` module into their own dedicated `:ui:play-integrity` and `:ui:key-attestation` modules.

Here is a summary of the changes:
- I moved all feature-specific source code (like Screens and ViewModels) and resources (such as drawables and strings) to their new modules.
- I updated the package names in all the files that were moved.
- I corrected all the import statements and resource `R` class references to match the new module structure.
- I relocated shared code (`Constants.kt`) and resources to the common `:ui` module so they can be accessed by multiple features.
- I refactored the `:ui:main` module to act as a navigation host, delegating the UI to the new feature modules.
- Finally, I verified the refactoring was complete with a successful `./gradlew assembleDebug` build.
