+++ android/ui/key-attestation/src/main/java/dev/keiji/deviceintegrity/ui/keyattestation/KeyAttestationViewModel.kt
@@ -1,4 +1,4 @@
-package dev.keiji.deviceintegrity.ui.main.keyattestation
+package dev.keiji.deviceintegrity.ui.keyattestation
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -18,7 +18,7 @@ import dev.keiji.deviceintegrity.repository.contract.KeyPairRepository
 import dev.keiji.deviceintegrity.repository.contract.KeyAttestationRepository
 import dev.keiji.deviceintegrity.repository.contract.exception.ServerException
 import dev.keiji.deviceintegrity.ui.common.InfoItem
-import dev.keiji.deviceintegrity.ui.main.common.KEY_ATTESTATION_DELAY_MS
+import dev.keiji.deviceintegrity.ui.common.KEY_ATTESTATION_DELAY_MS
 import dev.keiji.deviceintegrity.ui.common.ProgressConstants
 import dev.keiji.deviceintegrity.ui.util.Base64Utils
 import dev.keiji.deviceintegrity.ui.common.InfoItemFormatter
