+++ android/ui/key-attestation/src/main/java/dev/keiji/deviceintegrity/ui/keyattestation/ValueConverter.kt
@@ -1,4 +1,4 @@
-package dev.keiji.deviceintegrity.ui.main.keyattestation
+package dev.keiji.deviceintegrity.ui.keyattestation
 
 object ValueConverter {
