+++ android/ui/key-attestation/src/main/res/drawable/key_vertical_24dp_000000_fill1_wght400_grad0_opsz24.xml
