+++ android/ui/key-attestation/src/main/res/values-ja/strings.xml
@@ -1,0 +1,11 @@
+<resources>
+    <string name="key_attestation_result_label" >構成証明書の検証結果</string>
+
+    <string name="key_attestation_title_step1">Step 1. 鍵のアルゴリズムを選択</string>
+    <string name="key_attestation_title_step2">Step 2. サーバーからNonce/Challengeを取得</string>
+    <string name="key_attestation_title_step2_ecdh">Step 2. サーバーからNonce/Challenge/PublicKeyを取得</string>
+    <string name="key_attestation_title_step3">Step 3. キーペア（構成証明付き）を生成</string>
+    <string name="key_attestation_checkbox_label_strongbox">StrongBoxで鍵を生成する</string>
+    <string name="key_attestation_title_step4">Step 4. キーペアと構成証明を検証</string>
+    <string name="key_attestation_unsupported">この端末はKey Attestation（構成証明）をサポートしていません</string>
+</resources>
