+++ android/ui/main/src/main/java/dev/keiji/deviceintegrity/ui/main/AppScreen.kt
@@ -2,14 +2,28 @@ package dev.keiji.deviceintegrity.ui.main
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
-import dev.keiji.deviceintegrity.ui.main.R
 
 sealed class AppScreen(
     val route: String,
     @StringRes val label: Int,
     @param:DrawableRes val icon: Int
 ) {
-    object PlayIntegrity : AppScreen("play_integrity", R.string.app_screen_play_integrity_label, R.drawable.ic_shield)
-    object KeyAttestation : AppScreen("key_attestation", R.string.app_screen_key_attestation_label, R.drawable.key_vertical_24dp_000000_fill1_wght400_grad0_opsz24)
-    object Menu : AppScreen("menu", R.string.app_screen_menu_label, R.drawable.ic_menu)
+    object PlayIntegrity : AppScreen(
+        "play_integrity",
+        dev.keiji.deviceintegrity.ui.playintegrity.R.string.app_screen_play_integrity_label,
+        dev.keiji.deviceintegrity.ui.R.drawable.ic_shield
+    )
+
+    object KeyAttestation : AppScreen(
+        "key_attestation",
+        dev.keiji.deviceintegrity.ui.keyattestation.R.string.app_screen_key_attestation_label,
+        dev.keiji.deviceintegrity.ui.keyattestation.R.drawable.key_vertical_24dp_000000_fill1_wght400_grad0_opsz24
+    )
+
+    object Menu :
+        AppScreen(
+            "menu",
+            dev.keiji.deviceintegrity.ui.R.string.app_screen_menu_label,
+            R.drawable.ic_menu
+        )
 }
