+++ android/ui/play-integrity/src/main/res/drawable/gpp_bad_48dp_000000_fill0_wght400_grad0_opsz48.xml
