+++ android/ui/play-integrity/src/main/res/drawable/shield_48dp_000000_fill1_wght400_grad0_opsz48.xml
