+++ android/ui/play-integrity/src/main/res/values/strings.xml
@@ -1,0 +1,21 @@
+<resources>
+    <string name="app_screen_play_integrity_label" translatable="false">Play Integrity</string>
+
+    <string name="classic_pi_title_step1">Step 1. Fetch Nonce from Server</string>
+    <string name="classic_pi_button_fetch_nonce" translatable="false">Fetch Nonce</string>
+    <string name="classic_pi_label_nonce" translatable="false">Nonce: %s</string>
+    <string name="classic_pi_title_step2">Step 2. Request Token</string>
+    <string name="classic_pi_button_request_integrity_token" translatable="false">Request Integrity Token</string>
+    <string name="classic_pi_title_step3">Step 3. Verify Token</string>
+    <string name="classic_pi_button_request_verify_token" translatable="false">Request Verify Token</string>
+    <string name="classic_pi_label_error" translatable="false">Error: %s</string>
+
+    <string name="standard_pi_title_step1">Step 1. Set Content to Verify</string>
+    <string name="standard_pi_label_content" translatable="false">Content</string>
+    <string name="standard_pi_title_step2">Step 2. Request Token</string>
+    <string name="standard_pi_button_request_integrity_token" translatable="false">Request Integrity Token</string>
+    <string name="standard_pi_label_request_hash" translatable="false">requestHash: %s</string>
+    <string name="standard_pi_title_step3">Step 3. Verify Token</string>
+    <string name="standard_pi_button_request_verify_token" translatable="false">Request Verify Token</string>
+    <string name="standard_pi_label_error" translatable="false">Error: %s</string>
+</resources>
