+++ android/ui/src/main/java/dev/keiji/deviceintegrity/ui/common/Constants.kt
@@ -1,4 +1,4 @@
-package dev.keiji.deviceintegrity.ui.main.common
+package dev.keiji.deviceintegrity.ui.common
 
 const val VERIFY_TOKEN_DELAY_MS = 5000L
 const val DEBUG_VERIFY_TOKEN_DELAY_MS = 2000L
