+++ android/ui/src/main/res/drawable/ic_shield.xml
