+++ android/ui/src/main/res/values-ja/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="copied_to_clipboard">クリップボードにコピーしました</string>
+</resources>
