+++ android/ui/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_screen_menu_label" translatable="false">Menu</string>
+    <string name="copied_to_clipboard">Copied to clipboard</string>
+</resources>
